Bug Fix: match app-specific preferences
the id of the app-specific preference must be identical to the solution id
in the solution registry (e.g com.microsoft.windows.screenResolution).
Actually it should be
http://registry.gpii.net/applications/com.microsoft.windows.screenResolution
to be constitent with the JSON-LD model of the Preference but this does
not work with the Cloud4all specification. In Cloud4all, the app-specific
preferences corresponds to the full URI but in the output, only the
solution id of the solution registry is required.